+++ apps/frontend-web/app/test/[id]/page.tsx
@@ -1,43 +1,13 @@
 'use client';
 
-import { TestInput, useApiClient } from '@test-assessment/cms-graphql-api';
-import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 import Modal from './component/modal';
-import { transformQuestionDuplicate } from './utils/helper';
+import useTestDetail from './utils';
 
 const Page = ({ params }) => {
   const [showModal, setShowModal] = useState(false);
   const testId = params["id"];
-  const { apiClient } = useApiClient()
-  const router = useRouter();
-
-  const onDuplicateTest = async () => {
-    try {
-      const res = await apiClient.getDetailTest({ id: testId });
-      const testDetail = res.test.data;
-      if (!testDetail) return alert("Test not found.");
-      const testDetailAtr = testDetail.attributes;
-      const newTest: TestInput = {
-        name: `${testDetailAtr.name} ( Copy )`,
-        passingScore: testDetailAtr.passingScore,
-        timeLimit: testDetailAtr.timeLimit,
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        questions: transformQuestionDuplicate(testDetailAtr.questions as any),
-        position: testDetailAtr.position.data.id,
-        level: testDetailAtr.level,
-        publishedAt: null
-      };
-      const resDuplicateTest = await apiClient.createTest({ data: newTest });
-      const testDuplicateId = resDuplicateTest.createTest.data.id;
-      if (testDuplicateId) return router.push(`/test/${testId}/edit`);
-      alert("Duplicate fail, pls check again.");
-    } catch (err) {
-      const errors = err?.response?.errors || [];
-      const messErrors = errors.map(item => item.message);
-      if (messErrors.length > 0) return alert(messErrors[0])
-    }
-  };
+  const { onDuplicateTest } = useTestDetail({ testId });
 
   return (
     <>
